move web app commands
