+++ extensions/positron-python/package.json
@@ -427,42 +427,6 @@
                 "category": "Python",
                 "command": "python.installJupyter",
                 "title": "%python.command.python.installJupyter.title%"
-            },
-            {
-                "category": "Python",
-                "command": "python.runDashApp",
-                "title": "%python.command.python.runDashApp.title%",
-                "enablement": "config.python.enableWebAppSupport"
-            },
-            {
-                "category": "Python",
-                "command": "python.runFastAPIApp",
-                "title": "%python.command.python.runFastAPIApp.title%",
-                "enablement": "config.python.enableWebAppSupport"
-            },
-            {
-                "category": "Python",
-                "command": "python.runFlaskApp",
-                "title": "%python.command.python.runFlaskApp.title%",
-                "enablement": "config.python.enableWebAppSupport"
-            },
-            {
-                "category": "Python",
-                "command": "python.runGradioApp",
-                "title": "%python.command.python.runGradioApp.title%",
-                "enablement": "config.python.enableWebAppSupport"
-            },
-            {
-                "category": "Python",
-                "command": "python.runShinyApp",
-                "title": "%python.command.python.runShinyApp.title%",
-                "enablement": "config.python.enableWebAppSupport"
-            },
-            {
-                "category": "Python",
-                "command": "python.runStreamlitApp",
-                "title": "%python.command.python.runStreamlitApp.title%",
-                "enablement": "config.python.enableWebAppSupport"
             }
         ],
         "configuration": {
@@ -518,9 +482,9 @@
                     "scope": "application",
                     "type": "boolean"
                 },
-                "python.enableWebAppSupportSupport": {
+                "python.enableWebAppSupport": {
                     "defaut": false,
-                    "description": "%python.enableWebAppSupportSupport.description%",
+                    "description": "%python.enableWebAppSupport.description%",
                     "scope": "application",
                     "type": "boolean"
                 },
@@ -1260,7 +1224,7 @@
                 "command": "python.execInConsole",
                 "key": "ctrl+shift+enter",
                 "mac": "cmd+shift+enter",
-                "when": "editorTextFocus && editorLangId == python && !findInputFocussed && !replaceInputFocussed && !jupyter.ownsSelection && !notebookEditorFocused"
+                "when": "editorTextFocus && editorLangId == python && !findInputFocussed && !replaceInputFocussed && !jupyter.ownsSelection && !notebookEditorFocused && !(config.python.enableWebAppSupport && pythonAppFramework)"
             },
             {
                 "command": "python.execSelectionInConsole",
@@ -1400,7 +1364,7 @@
                     "category": "Python",
                     "command": "python.execInConsole",
                     "title": "%python.command.python.execInConsole.title%",
-                    "when": "!virtualWorkspace && shellExecutionSupported && editorLangId == python"
+                    "when": "!virtualWorkspace && shellExecutionSupported && editorLangId == python && !(config.python.enableWebAppSupport && pythonAppFramework)"
                 },
                 {
                     "category": "Python",
@@ -1526,7 +1490,7 @@
                 {
                     "command": "python.execInConsole",
                     "group": "Python",
-                    "when": "resourceLangId == python && !virtualWorkspace && shellExecutionSupported"
+                    "when": "resourceLangId == python && !virtualWorkspace && shellExecutionSupported && !(config.python.enableWebAppSupport && pythonAppFramework)"
                 },
                 {
                     "command": "python.execInTerminal",
@@ -1561,13 +1525,13 @@
                     "command": "python.execInConsole",
                     "group": "navigation@0",
                     "title": "%python.command.python.execInConsole.title%",
-                    "when": "resourceLangId == python && !isInDiffEditor && !virtualWorkspace && shellExecutionSupported"
+                    "when": "resourceLangId == python && !isInDiffEditor && !virtualWorkspace && shellExecutionSupported && !(config.python.enableWebAppSupport && pythonAppFramework)"
                 },
                 {
                     "command": "python.execDashInTerminal",
                     "group": "navigation@1",
                     "title": "%python.command.python.execDashInTerminal.title%",
-                    "when": "pythonAppFramework == dash && config.python.enableWebAppSupportSupport"
+                    "when": "pythonAppFramework == dash && config.python.enableWebAppSupport"
                 },
                 {
                     "command": "python.execGradioInTerminal",
